Move sub-classes of "ColumnHeader" and "ErrorProvider" classes to their own files
