+++ src/System.Windows.Forms/src/System/Windows/Forms/ErrorProvider.cs
@@ -812,890 +812,5 @@ namespace System.Windows.Forms
         }
 
         private bool ShouldSerializeIcon() => Icon != DefaultIcon;
-
-        /// <summary>
-        ///  There is one ErrorWindow for each control parent. It is parented to the
-        ///  control parent. The window's region is made up of the regions from icons
-        ///  of all child icons. The window's size is the enclosing rectangle for all
-        ///  the regions. A tooltip window is created as a child of this window. The
-        ///  rectangle associated with each error icon being displayed is added as a
-        ///  tool to the tooltip window.
-        /// </summary>
-        internal class ErrorWindow : NativeWindow
-        {
-            private static readonly int s_accessibilityProperty = PropertyStore.CreateKey();
-
-            private readonly List<ControlItem> _items = new List<ControlItem>();
-            private readonly Control _parent;
-            private readonly ErrorProvider _provider;
-            private Rectangle _windowBounds;
-            private Timer _timer;
-            private NativeWindow _tipWindow;
-
-            /// <summary>
-            ///  Construct an error window for this provider and control parent.
-            /// </summary>
-            public ErrorWindow(ErrorProvider provider, Control parent)
-            {
-                _provider = provider;
-                _parent = parent;
-                Properties = new PropertyStore();
-            }
-
-            /// <summary>
-            ///  The Accessibility Object for this ErrorProvider
-            /// </summary>
-            internal AccessibleObject AccessibilityObject
-            {
-                get
-                {
-                    AccessibleObject accessibleObject = (AccessibleObject)Properties.GetObject(s_accessibilityProperty);
-
-                    if (accessibleObject is null)
-                    {
-                        accessibleObject = CreateAccessibilityInstance();
-                        Properties.SetObject(s_accessibilityProperty, accessibleObject);
-                    }
-
-                    return accessibleObject;
-                }
-            }
-
-            /// <summary>
-            ///  This is called when a control would like to show an error icon.
-            /// </summary>
-            public void Add(ControlItem item)
-            {
-                _items.Add(item);
-                if (!EnsureCreated())
-                {
-                    return;
-                }
-
-                var toolInfo = new ComCtl32.ToolInfoWrapper<ErrorWindow>(this, item.Id, ComCtl32.TTF.SUBCLASS, item.Error);
-                toolInfo.SendMessage(_tipWindow, (User32.WM)ComCtl32.TTM.ADDTOOLW);
-
-                Update(timerCaused: false);
-            }
-
-            internal List<ControlItem> ControlItems => _items;
-
-            /// <summary>
-            ///  Constructs the new instance of the accessibility object for this ErrorProvider. Subclasses
-            ///  should not call base.CreateAccessibilityObject.
-            /// </summary>
-            private AccessibleObject CreateAccessibilityInstance()
-            {
-                return new ErrorWindowAccessibleObject(this);
-            }
-
-            /// <summary>
-            ///  Called to get rid of any resources the Object may have.
-            /// </summary>
-            public void Dispose() => EnsureDestroyed();
-
-            /// <summary>
-            ///  Make sure the error window is created, and the tooltip window is created.
-            /// </summary>
-            bool EnsureCreated()
-            {
-                if (Handle == IntPtr.Zero)
-                {
-                    if (!_parent.IsHandleCreated)
-                    {
-                        return false;
-                    }
-                    CreateParams cparams = new CreateParams
-                    {
-                        Caption = string.Empty,
-                        Style = (int)(User32.WS.VISIBLE | User32.WS.CHILD),
-                        ClassStyle = (int)User32.CS.DBLCLKS,
-                        X = 0,
-                        Y = 0,
-                        Width = 0,
-                        Height = 0,
-                        Parent = _parent.Handle
-                    };
-
-                    CreateHandle(cparams);
-
-                    var icc = new ComCtl32.INITCOMMONCONTROLSEX
-                    {
-                        dwICC = ComCtl32.ICC.TAB_CLASSES
-                    };
-                    ComCtl32.InitCommonControlsEx(ref icc);
-
-                    cparams = new CreateParams
-                    {
-                        Parent = Handle,
-                        ClassName = ComCtl32.WindowClasses.TOOLTIPS_CLASS,
-                        Style = (int)ComCtl32.TTS.ALWAYSTIP
-                    };
-                    _tipWindow = new NativeWindow();
-                    _tipWindow.CreateHandle(cparams);
-
-                    User32.SendMessageW(_tipWindow, (User32.WM)ComCtl32.TTM.SETMAXTIPWIDTH, IntPtr.Zero, (IntPtr)SystemInformation.MaxWindowTrackSize.Width);
-                    User32.SetWindowPos(
-                        new HandleRef(_tipWindow, _tipWindow.Handle),
-                        User32.HWND_TOP,
-                        flags: User32.SWP.NOSIZE | User32.SWP.NOMOVE | User32.SWP.NOACTIVATE);
-                    User32.SendMessageW(_tipWindow, (User32.WM)ComCtl32.TTM.SETDELAYTIME, (IntPtr)ComCtl32.TTDT.INITIAL, (IntPtr)0);
-                }
-
-                return true;
-            }
-
-            /// <summary>
-            ///  Destroy the timer, toolwindow, and the error window itself.
-            /// </summary>
-            private void EnsureDestroyed()
-            {
-                if (_timer != null)
-                {
-                    _timer.Dispose();
-                    _timer = null;
-                }
-                if (_tipWindow != null)
-                {
-                    _tipWindow.DestroyHandle();
-                    _tipWindow = null;
-                }
-
-                // Hide the window and invalidate the parent to ensure
-                // that we leave no visual artifacts. given that we
-                // have a bizare region window, this is needed.
-                User32.SetWindowPos(
-                    new HandleRef(this, Handle),
-                    User32.HWND_TOP,
-                    _windowBounds.X,
-                    _windowBounds.Y,
-                    _windowBounds.Width,
-                    _windowBounds.Height,
-                    User32.SWP.HIDEWINDOW | User32.SWP.NOSIZE | User32.SWP.NOMOVE);
-                _parent?.Invalidate(true);
-                DestroyHandle();
-            }
-
-            private unsafe void MirrorDcIfNeeded(Gdi32.HDC hdc)
-            {
-                if (_parent.IsMirrored)
-                {
-                    // Mirror the DC
-                    Gdi32.SetMapMode(hdc, Gdi32.MM.ANISOTROPIC);
-                    Gdi32.GetViewportExtEx(hdc, out Size originalExtents);
-                    Gdi32.SetViewportExtEx(hdc, -originalExtents.Width, originalExtents.Height, null);
-                    Gdi32.GetViewportOrgEx(hdc, out Point originalOrigin);
-                    Gdi32.SetViewportOrgEx(hdc, originalOrigin.X + _windowBounds.Width - 1, originalOrigin.Y, null);
-                }
-            }
-
-            /// <summary>
-            ///  This is called when the error window needs to paint. We paint each icon at its correct location.
-            /// </summary>
-            private unsafe void OnPaint()
-            {
-                using var hdc = new User32.BeginPaintScope(Handle);
-                using var save = new Gdi32.SaveDcScope(hdc);
-
-                MirrorDcIfNeeded(hdc);
-
-                for (int i = 0; i < _items.Count; i++)
-                {
-                    ControlItem item = _items[i];
-                    Rectangle bounds = item.GetIconBounds(_provider.Region.Size);
-                    User32.DrawIconEx(
-                        hdc,
-                        bounds.X - _windowBounds.X,
-                        bounds.Y - _windowBounds.Y,
-                        _provider.Region,
-                        bounds.Width, bounds.Height);
-                }
-            }
-
-            protected override void OnThreadException(Exception e)
-            {
-                Application.OnThreadException(e);
-            }
-
-            /// <summary>
-            ///  This is called when an error icon is flashing, and the view needs to be updatd.
-            /// </summary>
-            private void OnTimer(object sender, EventArgs e)
-            {
-                int blinkPhase = 0;
-                for (int i = 0; i < _items.Count; i++)
-                {
-                    blinkPhase += _items[i].BlinkPhase;
-                }
-                if (blinkPhase == 0 && _provider.BlinkStyle != ErrorBlinkStyle.AlwaysBlink)
-                {
-                    Debug.Assert(_timer != null);
-                    _timer.Stop();
-                }
-                Update(timerCaused: true);
-            }
-
-            private void OnToolTipVisibilityChanging(IntPtr id, bool toolTipShown)
-            {
-                for (int i = 0; i < _items.Count; i++)
-                {
-                    if (_items[i].Id == id)
-                    {
-                        _items[i].ToolTipShown = toolTipShown;
-                    }
-                }
-#if DEBUG
-                int shownTooltips = 0;
-                for (int j = 0; j < _items.Count; j++)
-                {
-                    if (_items[j].ToolTipShown)
-                    {
-                        shownTooltips++;
-                    }
-                }
-                Debug.Assert(shownTooltips <= 1);
-#endif
-            }
-
-            /// <summary>
-            ///  Retrieves our internal property storage object. If you have a property
-            ///  whose value is not always set, you should store it in here to save
-            ///  space.
-            /// </summary>
-            internal PropertyStore Properties { get; }
-
-            /// <summary>
-            ///  This is called when a control no longer needs to display an error icon.
-            /// </summary>
-            public void Remove(ControlItem item)
-            {
-                _items.Remove(item);
-
-                if (_tipWindow != null)
-                {
-                    var info = new ComCtl32.ToolInfoWrapper<ErrorWindow>(this, item.Id);
-                    info.SendMessage(_tipWindow, (User32.WM)ComCtl32.TTM.DELTOOLW);
-                }
-
-                if (_items.Count == 0)
-                {
-                    EnsureDestroyed();
-                }
-                else
-                {
-                    Update(timerCaused: false);
-                }
-            }
-
-            /// <summary>
-            ///  Start the blinking process. The timer will fire until there are no more
-            ///  icons that need to blink.
-            /// </summary>
-            public void StartBlinking()
-            {
-                if (_timer is null)
-                {
-                    _timer = new Timer();
-                    _timer.Tick += new EventHandler(OnTimer);
-                }
-
-                _timer.Interval = _provider.BlinkRate;
-                _timer.Start();
-                Update(timerCaused: false);
-            }
-
-            public void StopBlinking()
-            {
-                _timer?.Stop();
-                Update(timerCaused: false);
-            }
-
-            /// <summary>
-            ///  Move and size the error window, compute and set the window region, set the tooltip
-            ///  rectangles and descriptions. This basically brings the error window up to date with
-            ///  the internal data structures.
-            /// </summary>
-            public unsafe void Update(bool timerCaused)
-            {
-                IconRegion iconRegion = _provider.Region;
-                Size size = iconRegion.Size;
-                _windowBounds = Rectangle.Empty;
-                for (int i = 0; i < _items.Count; i++)
-                {
-                    ControlItem item = _items[i];
-                    Rectangle iconBounds = item.GetIconBounds(size);
-                    if (_windowBounds.IsEmpty)
-                    {
-                        _windowBounds = iconBounds;
-                    }
-                    else
-                    {
-                        _windowBounds = Rectangle.Union(_windowBounds, iconBounds);
-                    }
-                }
-
-                using var windowRegion = new Region(new Rectangle(0, 0, 0, 0));
-
-                for (int i = 0; i < _items.Count; i++)
-                {
-                    ControlItem item = _items[i];
-                    Rectangle iconBounds = item.GetIconBounds(size);
-                    iconBounds.X -= _windowBounds.X;
-                    iconBounds.Y -= _windowBounds.Y;
-
-                    bool showIcon = true;
-                    if (!item.ToolTipShown)
-                    {
-                        switch (_provider.BlinkStyle)
-                        {
-                            case ErrorBlinkStyle.NeverBlink:
-                                // always show icon
-                                break;
-                            case ErrorBlinkStyle.BlinkIfDifferentError:
-                                showIcon = (item.BlinkPhase == 0) || (item.BlinkPhase > 0 && (item.BlinkPhase & 1) == (i & 1));
-                                break;
-                            case ErrorBlinkStyle.AlwaysBlink:
-                                showIcon = ((i & 1) == 0) == _provider._showIcon;
-                                break;
-                        }
-                    }
-
-                    if (showIcon)
-                    {
-                        iconRegion.Region.Translate(iconBounds.X, iconBounds.Y);
-                        windowRegion.Union(iconRegion.Region);
-                        iconRegion.Region.Translate(-iconBounds.X, -iconBounds.Y);
-                    }
-
-                    if (_tipWindow != null)
-                    {
-                        ComCtl32.TTF flags = ComCtl32.TTF.SUBCLASS;
-                        if (_provider.RightToLeft)
-                        {
-                            flags |= ComCtl32.TTF.RTLREADING;
-                        }
-
-                        var toolInfo = new ComCtl32.ToolInfoWrapper<ErrorWindow>(this, item.Id, flags, item.Error, iconBounds);
-                        toolInfo.SendMessage(_tipWindow, (User32.WM)ComCtl32.TTM.SETTOOLINFOW);
-                    }
-
-                    if (timerCaused && item.BlinkPhase > 0)
-                    {
-                        item.BlinkPhase--;
-                    }
-                }
-
-                if (timerCaused)
-                {
-                    _provider._showIcon = !_provider._showIcon;
-                }
-
-                using var hdc = new User32.GetDcScope(Handle);
-                using var save = new Gdi32.SaveDcScope(hdc);
-                MirrorDcIfNeeded(hdc);
-
-                using Graphics g = hdc.CreateGraphics();
-                using var windowRegionHandle = new Gdi32.RegionScope(windowRegion, g);
-                if (User32.SetWindowRgn(this, windowRegionHandle, BOOL.TRUE) != 0)
-                {
-                    // The HWnd owns the region.
-                    windowRegionHandle.RelinquishOwnership();
-                }
-
-                User32.SetWindowPos(
-                    new HandleRef(this, Handle),
-                    User32.HWND_TOP,
-                    _windowBounds.X,
-                    _windowBounds.Y,
-                    _windowBounds.Width,
-                    _windowBounds.Height,
-                    User32.SWP.NOACTIVATE);
-                User32.InvalidateRect(new HandleRef(this, Handle), null, BOOL.FALSE);
-            }
-
-            /// <summary>
-            ///  Handles the WM_GETOBJECT message. Used for accessibility.
-            /// </summary>
-            private void WmGetObject(ref Message m)
-            {
-                Debug.WriteLineIf(CompModSwitches.MSAA.TraceInfo, "In WmGetObject, this = " + GetType().FullName + ", lParam = " + m.LParam.ToString());
-
-                if (m.Msg == (int)User32.WM.GETOBJECT && m.LParam == (IntPtr)NativeMethods.UiaRootObjectId)
-                {
-                    // If the requested object identifier is UiaRootObjectId,
-                    // we should return an UI Automation provider using the UiaReturnRawElementProvider function.
-                    InternalAccessibleObject intAccessibleObject = new InternalAccessibleObject(AccessibilityObject);
-                    m.Result = UiaCore.UiaReturnRawElementProvider(
-                        new HandleRef(this, Handle),
-                        m.WParam,
-                        m.LParam,
-                        intAccessibleObject);
-
-                    return;
-                }
-
-                // some accessible object requested that we don't care about, so do default message processing
-                DefWndProc(ref m);
-            }
-
-            /// <summary>
-            ///  Called when the error window gets a windows message.
-            /// </summary>
-            protected unsafe override void WndProc(ref Message m)
-            {
-                switch ((User32.WM)m.Msg)
-                {
-                    case User32.WM.GETOBJECT:
-                        WmGetObject(ref m);
-                        break;
-                    case User32.WM.NOTIFY:
-                        User32.NMHDR* nmhdr = (User32.NMHDR*)m.LParam;
-                        if (nmhdr->code == (int)ComCtl32.TTN.SHOW || nmhdr->code == (int)ComCtl32.TTN.POP)
-                        {
-                            OnToolTipVisibilityChanging(nmhdr->idFrom, nmhdr->code == (int)ComCtl32.TTN.SHOW);
-                        }
-                        break;
-                    case User32.WM.ERASEBKGND:
-                        break;
-                    case User32.WM.PAINT:
-                        OnPaint();
-                        break;
-                    default:
-                        base.WndProc(ref m);
-                        break;
-                }
-            }
-        }
-
-        /// <summary>
-        ///  There is one ControlItem for each control that the ErrorProvider is tracking state for.
-        ///  It contains the values of all the extender properties.
-        /// </summary>
-        internal class ControlItem
-        {
-            private static readonly int s_accessibilityProperty = PropertyStore.CreateKey();
-
-            private string _error;
-            private readonly Control _control;
-            private ErrorWindow _window;
-            private readonly ErrorProvider _provider;
-            private int _iconPadding;
-            private ErrorIconAlignment _iconAlignment;
-            private const int _startingBlinkPhase = 10; // We want to blink 5 times
-
-            /// <summary>
-            ///  Construct the item with its associated control, provider, and a unique ID. The ID is
-            ///  used for the tooltip ID.
-            /// </summary>
-            public ControlItem(ErrorProvider provider, Control control, IntPtr id)
-            {
-                _iconAlignment = DefaultIconAlignment;
-                _error = string.Empty;
-                Id = id;
-                _control = control;
-                _provider = provider;
-                _control.HandleCreated += new EventHandler(OnCreateHandle);
-                _control.HandleDestroyed += new EventHandler(OnDestroyHandle);
-                _control.LocationChanged += new EventHandler(OnBoundsChanged);
-                _control.SizeChanged += new EventHandler(OnBoundsChanged);
-                _control.VisibleChanged += new EventHandler(OnParentVisibleChanged);
-                _control.ParentChanged += new EventHandler(OnParentVisibleChanged);
-                Properties = new PropertyStore();
-            }
-
-            /// <summary>
-            ///  The Accessibility Object for this ErrorProvider
-            /// </summary>
-            internal AccessibleObject AccessibilityObject
-            {
-                get
-                {
-                    AccessibleObject accessibleObject = (AccessibleObject)Properties.GetObject(s_accessibilityProperty);
-
-                    if (accessibleObject is null)
-                    {
-                        accessibleObject = CreateAccessibilityInstance();
-                        Properties.SetObject(s_accessibilityProperty, accessibleObject);
-                    }
-
-                    return accessibleObject;
-                }
-            }
-
-            /// <summary>
-            ///  Constructs the new instance of the accessibility object for this ErrorProvider. Subclasses
-            ///  should not call base.CreateAccessibilityObject.
-            /// </summary>
-            private AccessibleObject CreateAccessibilityInstance()
-            {
-                return new ControlItemAccessibleObject(this, _window, _control.ParentInternal, _provider);
-            }
-
-            public void Dispose()
-            {
-                if (_control != null)
-                {
-                    _control.HandleCreated -= new EventHandler(OnCreateHandle);
-                    _control.HandleDestroyed -= new EventHandler(OnDestroyHandle);
-                    _control.LocationChanged -= new EventHandler(OnBoundsChanged);
-                    _control.SizeChanged -= new EventHandler(OnBoundsChanged);
-                    _control.VisibleChanged -= new EventHandler(OnParentVisibleChanged);
-                    _control.ParentChanged -= new EventHandler(OnParentVisibleChanged);
-                }
-
-                _error = string.Empty;
-            }
-
-            /// <summary>
-            ///  Returns the unique ID for this control. The ID used as the tooltip ID.
-            /// </summary>
-            public IntPtr Id { get; }
-
-            /// <summary>
-            ///  Returns or set the phase of blinking that this control is currently
-            ///  in.  If zero, the control is not blinking. If odd, then the control
-            ///  is blinking, but invisible. If even, the control is blinking and
-            ///  currently visible. Each time the blink timer fires, this value is
-            ///  reduced by one (until zero), thus causing the error icon to appear
-            ///  or disappear.
-            /// </summary>
-            public int BlinkPhase { get; set; }
-
-            /// <summary>
-            ///  Returns or sets the icon padding for the control.
-            /// </summary>
-            public int IconPadding
-            {
-                get => _iconPadding;
-                set
-                {
-                    if (_iconPadding == value)
-                    {
-                        return;
-                    }
-
-                    _iconPadding = value;
-                    UpdateWindow();
-                }
-            }
-
-            /// <summary>
-            ///  Returns or sets the error description string for the control.
-            /// </summary>
-            public string Error
-            {
-                get => _error;
-                set
-                {
-                    if (value is null)
-                    {
-                        value = string.Empty;
-                    }
-
-                    // If the error is the same and the blinkStyle is not AlwaysBlink, then
-                    // we should not add the error and not start blinking.
-                    if (_error.Equals(value) && _provider.BlinkStyle != ErrorBlinkStyle.AlwaysBlink)
-                    {
-                        return;
-                    }
-
-                    bool adding = _error.Length == 0;
-                    _error = value;
-                    if (value.Length == 0)
-                    {
-                        RemoveFromWindow();
-                    }
-                    else
-                    {
-                        if (adding)
-                        {
-                            AddToWindow();
-                        }
-                        else
-                        {
-                            if (_provider.BlinkStyle != ErrorBlinkStyle.NeverBlink)
-                            {
-                                StartBlinking();
-                            }
-                            else
-                            {
-                                UpdateWindow();
-                            }
-                        }
-                    }
-                }
-            }
-
-            /// <summary>
-            ///  Returns or sets the location of the error icon for the control.
-            /// </summary>
-            public ErrorIconAlignment IconAlignment
-            {
-                get => _iconAlignment;
-                set
-                {
-                    SourceGenerated.EnumValidator.Validate(value);
-
-                    if (_iconAlignment == value)
-                    {
-                        return;
-                    }
-
-                    _iconAlignment = value;
-                    UpdateWindow();
-                }
-            }
-
-            /// <summary>
-            ///  Returns true if the tooltip for this control item is currently shown.
-            /// </summary>
-            public bool ToolTipShown { get; set; }
-
-            internal ErrorIconAlignment RTLTranslateIconAlignment(ErrorIconAlignment align)
-            {
-                if (_provider.RightToLeft)
-                {
-                    switch (align)
-                    {
-                        case ErrorIconAlignment.TopLeft:
-                            return ErrorIconAlignment.TopRight;
-                        case ErrorIconAlignment.MiddleLeft:
-                            return ErrorIconAlignment.MiddleRight;
-                        case ErrorIconAlignment.BottomLeft:
-                            return ErrorIconAlignment.BottomRight;
-                        case ErrorIconAlignment.TopRight:
-                            return ErrorIconAlignment.TopLeft;
-                        case ErrorIconAlignment.MiddleRight:
-                            return ErrorIconAlignment.MiddleLeft;
-                        case ErrorIconAlignment.BottomRight:
-                            return ErrorIconAlignment.BottomLeft;
-                        default:
-                            Debug.Fail("Unknown ErrorIconAlignment value");
-                            return align;
-                    }
-                }
-                else
-                {
-                    return align;
-                }
-            }
-
-            /// <summary>
-            ///  Returns the location of the icon in the same coordinate system as the control being
-            ///  extended. The size passed in is the size of the icon.
-            /// </summary>
-            internal Rectangle GetIconBounds(Size size)
-            {
-                int x = 0;
-                int y = 0;
-
-                switch (RTLTranslateIconAlignment(IconAlignment))
-                {
-                    case ErrorIconAlignment.TopLeft:
-                    case ErrorIconAlignment.MiddleLeft:
-                    case ErrorIconAlignment.BottomLeft:
-                        x = _control.Left - size.Width - _iconPadding;
-                        break;
-                    case ErrorIconAlignment.TopRight:
-                    case ErrorIconAlignment.MiddleRight:
-                    case ErrorIconAlignment.BottomRight:
-                        x = _control.Right + _iconPadding;
-                        break;
-                }
-
-                switch (IconAlignment)
-                {
-                    case ErrorIconAlignment.TopLeft:
-                    case ErrorIconAlignment.TopRight:
-                        y = _control.Top;
-                        break;
-                    case ErrorIconAlignment.MiddleLeft:
-                    case ErrorIconAlignment.MiddleRight:
-                        y = _control.Top + (_control.Height - size.Height) / 2;
-                        break;
-                    case ErrorIconAlignment.BottomLeft:
-                    case ErrorIconAlignment.BottomRight:
-                        y = _control.Bottom - size.Height;
-                        break;
-                }
-
-                return new Rectangle(x, y, size.Width, size.Height);
-            }
-
-            /// <summary>
-            ///  If this control's error icon has been added to the error window, then update the
-            ///  window state because some property has changed.
-            /// </summary>
-            private void UpdateWindow() => _window?.Update(timerCaused: false);
-
-            /// <summary>
-            ///  If this control's error icon has been added to the error window, then start blinking
-            ///  the error window.
-            /// </summary>
-            private void StartBlinking()
-            {
-                if (_window != null)
-                {
-                    BlinkPhase = _startingBlinkPhase;
-                    _window.StartBlinking();
-                }
-            }
-
-            /// <summary>
-            ///  Add this control's error icon to the error window.
-            /// </summary>
-            private void AddToWindow()
-            {
-                // if we are recreating the control, then add the control.
-                if (_window is null &&
-                    (_control.Created || _control.RecreatingHandle) &&
-                    _control.Visible && _control.ParentInternal != null &&
-                    _error.Length > 0)
-                {
-                    _window = _provider.EnsureErrorWindow(_control.ParentInternal);
-                    _window.Add(this);
-                    // Make sure that we blink if the style is set to AlwaysBlink or BlinkIfDifferrentError
-                    if (_provider.BlinkStyle != ErrorBlinkStyle.NeverBlink)
-                    {
-                        StartBlinking();
-                    }
-                }
-            }
-
-            /// <summary>
-            ///  Remove this control's error icon from the error window.
-            /// </summary>
-            private void RemoveFromWindow()
-            {
-                if (_window != null)
-                {
-                    _window.Remove(this);
-                    _window = null;
-                }
-            }
-
-            /// <summary>
-            ///  This is called when a property on the control is changed.
-            /// </summary>
-            private void OnBoundsChanged(object sender, EventArgs e) => UpdateWindow();
-
-            void OnParentVisibleChanged(object sender, EventArgs e)
-            {
-                BlinkPhase = 0;
-                RemoveFromWindow();
-                AddToWindow();
-            }
-
-            /// <summary>
-            ///  Retrieves our internal property storage object. If you have a property
-            ///  whose value is not always set, you should store it in here to save
-            ///  space.
-            /// </summary>
-            private PropertyStore Properties { get; }
-
-            /// <summary>
-            ///  This is called when the control's handle is created.
-            /// </summary>
-            private void OnCreateHandle(object sender, EventArgs e) => AddToWindow();
-
-            /// <summary>
-            ///  This is called when the control's handle is destroyed.
-            /// </summary>
-            private void OnDestroyHandle(object sender, EventArgs e) => RemoveFromWindow();
-        }
-
-        /// <summary>
-        ///  This represents the HRGN of icon. The region is calculate from the icon's mask.
-        /// </summary>
-        internal class IconRegion : IHandle
-        {
-            private Region region;
-            private readonly Icon icon;
-
-            /// <summary>
-            ///  Constructor that takes an Icon and extracts its 16x16 version.
-            /// </summary>
-            public IconRegion(Icon icon)
-            {
-                this.icon = new Icon(icon, 16, 16);
-            }
-
-            /// <summary>
-            ///  Returns the handle of the icon.
-            /// </summary>
-            public IntPtr Handle => icon.Handle;
-
-            /// <summary>
-            ///  Returns the handle of the region.
-            /// </summary>
-            public Region Region
-            {
-                get
-                {
-                    if (region is null)
-                    {
-                        region = new Region(new Rectangle(0, 0, 0, 0));
-
-                        IntPtr mask = IntPtr.Zero;
-                        try
-                        {
-                            Size size = icon.Size;
-                            Bitmap bitmap = icon.ToBitmap();
-                            mask = ControlPaint.CreateHBitmapTransparencyMask(bitmap);
-                            bitmap.Dispose();
-
-                            // It is been observed that users can use non standard size icons (not a 16 bit multiples for width and height)
-                            // and GetBitmapBits method allocate bytes in multiple of 16 bits for each row. Following calculation is to get right width in bytes.
-                            int bitmapBitsAllocationSize = 16;
-
-                            // If width is not multiple of 16, we need to allocate BitmapBitsAllocationSize for remaining bits.
-                            int widthInBytes = 2 * ((size.Width + 15) / bitmapBitsAllocationSize); // its in bytes.
-                            byte[] bits = new byte[widthInBytes * size.Height];
-                            Gdi32.GetBitmapBits(mask, bits.Length, bits);
-
-                            for (int y = 0; y < size.Height; y++)
-                            {
-                                for (int x = 0; x < size.Width; x++)
-                                {
-                                    // see if bit is set in mask. bits in byte are reversed. 0 is black (set).
-                                    if ((bits[y * widthInBytes + x / 8] & (1 << (7 - (x % 8)))) == 0)
-                                    {
-                                        region.Union(new Rectangle(x, y, 1, 1));
-                                    }
-                                }
-                            }
-                            region.Intersect(new Rectangle(0, 0, size.Width, size.Height));
-                        }
-                        finally
-                        {
-                            if (mask != IntPtr.Zero)
-                            {
-                                Gdi32.DeleteObject(mask);
-                            }
-                        }
-                    }
-
-                    return region;
-                }
-            }
-
-            /// <summary>
-            ///  Return the size of the icon.
-            /// </summary>
-            public Size Size => icon.Size;
-
-            /// <summary>
-            ///  Release any resources held by this Object.
-            /// </summary>
-            public void Dispose()
-            {
-                if (region != null)
-                {
-                    region.Dispose();
-                    region = null;
-                }
-                icon.Dispose();
-            }
-        }
     }
 }
